Move inherited database files
