+++ .example.env
@@ -2,3 +2,7 @@ AWS_ACCESS_KEY_ID="/***AWS ACCESS KEY ID***/"
 AWS_SECRET_ACCESS_KEY="/***AWS SECRET ACCESS KEY***/"
 AWS_S3_BUCKET="/***S3 BUCKET NAME**/"
 AWS_BUCKET_BASE_URL="/***BUCKET BASE URL***/"
+
+CB_USERNAME="/***COUCHBASE USERNAME***/"
+CB_PASSWORD="/***COUCHBASE PASSWORD***/"
+CB_URI="/***COUCHBASE URI***/"
