+++ database/inheritedDatabase/PhotoModel.js
