+++ database/inheritedDatabase/index.js
